add PacketFrame class

+++ checksum.h
@@ -1,10 +1,10 @@
 #ifndef _CHECKSUM_H
 #define _CHECKSUM_H
 
-char checksum(char * data, unsigned short size) {
+char checksum(char * data, int size) {
     char checksum_result = 0;
 
-    for(unsigned short i = 0; i < size; i++) {
+    for(int i = 0; i < size; i++) {
         checksum_result += data[i];
     }
